Refactor the build process [1/2].
The data located in this module comes from, or is associated with,
one of two main sources: The zsh package itself, and the bits of glue
that make it work under magisk (Makefile, installer script, android specific
configs, etc).  Instead of lumping them together and committing random
prebuilt zsh stuff, the two are now separated until its time to build
the actual module, when one is overlaid on the other.